feat(menu): util component
